+++ refl1d/main.py
@@ -1,7 +1,13 @@
 #!/usr/bin/env python
 """
-Run bumps with refl1d plugin
+Run bumps with refl1d plugin.
+
+The *align* subcommand can be used on a completed DREAM fit to redraw the
+profile contours aligned to a different layer boundary.
+See :func:`refl1d.errors.run_errors` for details.
 """
+import sys
+
 from . import __version__
 
 def setup_bumps():
@@ -28,8 +34,15 @@ def cli():
     Install the Refl1D plugin into bumps and run the command line interface.
     """
     setup_bumps()
-    import bumps.cli
-    bumps.cli.main()
+
+    # TODO: Add subcommand support to bumps.
+    if len(sys.argv) > 2 and sys.argv[1] == 'align':
+        from .errors import run_errors
+        del sys.argv[1]
+        run_errors()
+    else:
+        import bumps.cli
+        bumps.cli.main()
 
 def gui():
     """