+++ CHANGES.rst
@@ -7,7 +7,7 @@ Note: updated docs now found at `<http://refl1d.readthedocs.org>`_
 2020-..-.. v0.8.12
 ==================
 * add irho to the profile contour plots
-* simplify realignment usage to 'refl1d -m refl1d.align ...'
+* simplify realignment usage to 'refl1d align ...'
 * reenable python 2.7 support
 
 2020-06-11 v0.8.11